#10072: Merge hardswish_bw to TTNN
